pyproject.toml
[project]
name = "the-trufiti-group"
version = "0.1.0"
description = ""
authors = [
    { name = "Finn Eilmann" },
    { name = "Till Flotow" },
    { name = "Hoang Trung Nguyen" },
]
requires-python = ">=3.12,<3.15"
readme = "README.md"
dependencies = [
    "polars (>=1.34.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "torch>=2.9.0,<3.0.0",
    "torchvision>=0.24.0,<0.25.0",
    "notebook (>=7.4.7,<8.0.0)",
    "tokenizers (>=0.22.1,<0.23.0)",
    "pyyaml>=6.0.0,<7.0.0",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
    "numpy>=1.26.4",
    "lingua-language-detector>=2.1.1",
    "torchtune>=0.6.1",
    "torchao>=0.14.1",
    "matplotlib>=3.10.7",
    "evaluate>=0.4.6",
    "bert-score>=0.3.13",
    "rouge-score>=0.1.2",
    "fastapi>=0.124.4",
    "uvicorn>=0.38.0",
    "pyspellchecker>=0.8.4",
    "language-tool-python>=3.2.0",
    "mauve-text>=0.4.0",
    "hf-xet>=1.2.0",
    "detoxify>=0.5.2",
    "lmstudio>=1.5.0",
    "openai>=2.14.0",
    "scipy-stubs==1.17.0.1",
    "emoji>=2.15.0",
    "deprecated>=1.3.1",
    "types-deprecated>=1.3.1.20251101",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pre-commit>=4.4.0,<5",
    "ruff>=0.14.5,<0.15",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
#    "I",   # isort
    "UP",  # pyupgrade (modernere Python-Syntax)
    "B",   # flake8-bugbear (findet häufige Bugs)
]
ignore = []


[tool.mypy]
[[tool.mypy.overrides]]
module = ["tokenizers", "tokenizers.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
